EmailSystem/requirements.txt
asgiref==3.6.0
Django==4.1.7
djangorestframework==3.14.0
pika==1.3.1
psycopg2==2.9.5
python-http-client==3.3.7
pytz==2022.7.1
sendgrid==6.10.0
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
